simplifies the abstract example
