+++ src/command/001_abstract/Command/Receiver.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace Project
+{
+    interface IReceiver
+    {
+        void Execute();
+    }
+
+    class Receiver : IReceiver
+    {
+        public void Execute()
+        {
+            Console.WriteLine($"{nameof(Receiver)} executed!");
+        }
+    }
+}
