--- src/command/001_abstract/Commands/ICommand.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Project
-{
-    interface ICommand
-    {
-        bool Execute(bool param1, int param2, string param3);
-    }
-}
