--- src/command/001_abstract/Receivers/IReceiver.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Project
-{
-    interface IReceiver
-    {
-        bool Execute(bool param1, int param2, string param3);
-    }
-}
